sketch: specifying counted wallet
Now I need to think how to deal with the ongoing calls.
Maybe I just need to specify that everything ongoing calls return at a particular place.

For this kind of thing, I need to generalize things a bit.  A predicate for a concrete state suffices.  Currently an equality is required.